requirements.txt
# Core dependencies
numpy>=1.19.0

# Optional GPU support (comment out if not needed)
cupy-cuda12x>=12.0.0  # For CUDA 12.x (alternative)

# Visualization
matplotlib>=3.3.0

# Optional: For examples and testing
scikit-learn>=0.24.0

# Optional: For Jupyter notebook examples
jupyter>=1.0.0
ipykernel>=6.0.0

# Development dependencies (optional)
pytest>=6.0.0